GameManagerのタイム管理をTimeManagerに移しました。
