+++ Assets/Material/AquaBall.mat
@@ -96,7 +96,7 @@ Material:
     - _WorkflowMode: 1
     - _ZWrite: 0
     m_Colors:
-    - _BaseColor: {r: 1, g: 1, b: 1, a: 0.14901961}
+    - _BaseColor: {r: 1, g: 1, b: 1, a: 0.78431374}
     - _Color: {r: 1, g: 1, b: 1, a: 0.14901961}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.2, g: 0.2, b: 0.2, a: 1}
