+++ Assets/Prefab/StartCube.prefab
@@ -106,3 +106,17 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5dbe24d5046fa42a3920101be8446b20, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  m_startExitPlayerEvent:
+    m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 11500000, guid: 3a84221bfc5654443ab9d7a8412f5d34, type: 3}
+        m_MethodName: 
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: 
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
