+++ Assets/Script/TimeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c859afee307d640b4bb77a2a2941fb3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
