+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/TestMapping.unity
-    guid: f6e40ca10933f4ee68bd63c545a94ca9
+    path: Assets/Scenes/GameScene.unity
+    guid: ab729dbffc7df4fcea5f8950494a5b50
   m_configObjects: {}
